requirements.txt

pylint==2.3.1
python-dotenv==0.10.1
jsonfield==2.0.2

Django==2.2.3
django-choices==1.7.0
django-admin-rangefilter==0.4.0
django-cors-headers==2.5.2
django-polymorphic==2.1.2
django-rest-polymorphic==0.1.8
django-timezone-field==3.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
django-modeltranslation==0.14.2


chardet==3.0.4
Django==2.2.3
django-admin-rangefilter==0.4.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-choices==1.7.0
django-cors-headers==2.5.2
django-modeltranslation==0.14.2
django-polymorphic==2.1.2
django-rest-polymorphic==0.1.8
django-timezone-field==3.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0

httpie==2.2.0
jsonfield==2.0.2
oauth2-provider==0.0
oauthlib==3.1.0
Pygments==2.6.1
PyJWT==1.7.1
pylint-django==2.0.15
requests==2.24.0
sqlparse==0.3.1



